+++ lib/crypt_signature.dart
@@ -6,6 +6,6 @@ export 'src/models/certificate.dart' show Certificate;
 export 'src/models/digest_result.dart' show DigestResult;
 export 'src/models/license.dart' show License;
 export 'src/models/pkcs7.dart' show PKCS7;
-export 'src/models/sign_request.dart' show MessageSignRequest, PKCS7MessageSignRequest, PKCS7HASHSignRequest, CustomSignRequest;
-export 'src/models/sign_result.dart' show SignResult, PKCS7SignResult;
+export 'src/models/sign_request.dart' show MessageSignRequest, CMSMessageSignRequest, CMSHashSignRequest, CustomSignRequest;
+export 'src/models/sign_result.dart' show SignResult, CMSSignResult;
 export 'src/utils/crypt_signature_theme.dart' show CryptSignatureTheme;