+++ lib/src/crypt_signature_api.dart
@@ -30,9 +30,8 @@ class CryptSignature {
 
   CryptSignature._(this.certificateRepository);
 
-  /// Открыть экран выбора сертификата и подписать
-  /// В случает подписи PKCS7 на Android возвращает [PKCS7SignResult]
-  /// Возвращает [SignResult] для всеъ остальных
+  /// Открыть экран выбора сертификата и подписать сообщение или хэш.
+  /// В случает подписи CMS возвращает [CMSSignResult].
   Future<T?> interface<T extends SignResult>(BuildContext context, SignRequest<T> signRequest, {CryptSignatureTheme? theme}) async {
     NavigatorState navigator = Navigator.of(context);
     Directory directory = await getApplicationDocumentsDirectory();