+++ lib/src/models/sign_result.dart
@@ -1,11 +1,8 @@
-import 'dart:convert';
-import 'dart:io';
-
 import 'package:crypt_signature/src/models/certificate.dart';
 import 'package:crypt_signature/src/models/pkcs7.dart';
 import 'package:crypt_signature/src/utils/extensions/extensions.dart';
 
-/// Результат подписи
+/// Результат подписи.
 class SignResult {
   /// Сертификат из контейнера с приватным ключем
   final Certificate certificate;
@@ -19,9 +16,7 @@ class SignResult {
   /// Алгоритм сигнатуры
   final String signatureAlgorithm;
 
-  SignResult(this.certificate, {required this.digest, required String signature, required this.signatureAlgorithm})
-      : // Нативные функции win32 возвращают развернутую сигнатуру
-        signature = Platform.isIOS ? reverseSignature(signature) : signature;
+  SignResult(this.certificate, {required this.digest, required this.signature, required this.signatureAlgorithm});
 
   SignResult.from(SignResult other)
       : certificate = other.certificate,
@@ -40,17 +35,15 @@ class SignResult {
 
     return stringBuffer.toString();
   }
-
-  static String reverseSignature(String signature) => base64.encode(base64.decode(signature.replaceAll("\n", "")).reversed.toList());
 }
 
-class PKCS7SignResult extends SignResult {
+class CMSSignResult extends SignResult {
   final PKCS7 pkcs7;
 
-  /// Хэш от изначального сообщения. Так как в signResult лежит хэш от атрибутов подписи
+  /// Хэш от изначального сообщения. Так как в [digest] лежит хэш от атрибутов подписи.
   final String initialDigest;
 
-  PKCS7SignResult(
+  CMSSignResult(
     this.pkcs7,
     super.certificate, {
     required this.initialDigest,
@@ -59,7 +52,7 @@ class PKCS7SignResult extends SignResult {
     required super.signatureAlgorithm,
   });
 
-  PKCS7SignResult.from(this.pkcs7, SignResult signResult, {required this.initialDigest}) : super.from(signResult);
+  CMSSignResult.from(this.pkcs7, SignResult signResult, {required this.initialDigest}) : super.from(signResult);
 
   @override
   String toString() => "${super.toString()}\nPKCS7: $pkcs7\nInitialDigest: $initialDigest";