+++ lib/src/utils/exceptions/api_response_exception.dart
@@ -3,4 +3,7 @@ class ApiResponseException implements Exception {
   final String? details;
 
   ApiResponseException(this.message, this.details);
+
+  @override
+  String toString() => "ApiResponseException. $message $details";
 }